word definition service, text service tests halfway done

+++ MichtavaSol/Services/SubjectService.cs
@@ -22,7 +22,7 @@
         }
 
 
-             public Subject GetByName(string subjectName)
+        public Subject GetByName(string subjectName)
         {
             return this.subjectRepository.GetByName(subjectName);
         }
@@ -34,6 +34,11 @@
 
         public MichtavaResult Add(Subject subject)
         {
+            if (subjectRepository.GetByName(subject.Name) != null)
+                return new MichtavaFailure("כבר קיים נושא עם אותו השם");
+
+
+
             this.subjectRepository.Add(subject);
             this.subjectRepository.SaveChanges();
             return new MichtavaSuccess();
@@ -42,25 +47,56 @@
 
         public MichtavaResult Update(Subject subject)
         {
-            this.subjectRepository.Update(subject);
-            this.subjectRepository.SaveChanges();
-            return new MichtavaSuccess();
+            if (this.GetById(subject.Id) == null)
+                return new MichtavaFailure("הנושא לא נמצא במערכת");
 
+            if (subjectRepository.Get(s => s.Name == subject.Name).Count() > 1)
+                return new MichtavaFailure();
+
+            if (subjectRepository.Get(s => s.Name == subject.Name).Count() == 1)
+                return new MichtavaFailure("כבר קיים נושא במערכת עם אותו השם");
+
+            if (subjectRepository.Get(s => s.Name == subject.Name).Count() == 0)
+            {
+                this.subjectRepository.Update(subject);
+                this.subjectRepository.SaveChanges();
+                return new MichtavaSuccess("נושא עודכן בהצלחה");
+            }
+            else
+                return new MichtavaFailure("טעות לא צפויה התרחשה");
         }
 
         public MichtavaResult Delete(Subject subject)
         {
+            Subject existing = this.subjectRepository.All().Where(y => y.Id == subject.Id).FirstOrDefault();
+
+
+            if (existing == null)
+                return new MichtavaFailure("הנושא לא נמצא במערכת");
+
+            //delete from all texts and scs...???
+
             this.subjectRepository.Delete(subject);
             this.subjectRepository.SaveChanges();
-            return new MichtavaSuccess();
+            return new MichtavaSuccess("הנושא נמחק בהצלחה");
 
         }
 
         public MichtavaResult HardDelete(Subject subject)
         {
+            Subject existing = this.subjectRepository.AllWithDeleted().Where(y => y.Id == subject.Id).FirstOrDefault();
+
+
+            if (existing == null)
+                return new MichtavaFailure("הנושא לא נמצא במערכת");
+
+            //delete from all ss and scs...???
+
+
             this.subjectRepository.HardDelete(subject);
             this.subjectRepository.SaveChanges();
-            return new MichtavaSuccess();
+            return new MichtavaSuccess("הנושא נמחק בהצלחה");
+
         }
     }
 }
